+++ lib/ui/widgets/tool_view.dart
@@ -1,33 +1,5 @@
 import 'package:flutter/material.dart';
 
-class ToolView extends StatelessWidget {
-  const ToolView({
-    super.key,
-    required this.title,
-    required this.content,
-    this.actions,
-    this.fab,
-  });
-  final Widget title;
-  final Widget content;
-  final List<Widget>? actions;
-  final Widget? fab;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: title,
-        centerTitle: false,
-        actions: actions,
-      ),
-      floatingActionButton: fab,
-      body: content,
-      backgroundColor: Theme.of(context).colorScheme.background,
-    );
-  }
-}
-
 @Deprecated('use ScrollableScaffold')
 class ScrollableToolView extends StatelessWidget {
   const ScrollableToolView({