+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - blurhash to image.
 - json/yaml converter add formater.
 - add more hash algorithms.
+- new common converter
 
 ### Changed
 - Material 3 Support.